Chore: Crear package de los modulos de cdp (#30)
* Crear paquete con los modulos de cdp

* Solucionar argumento faltante al crear Usuarios
