+++ src/cdp/settings/motor_data.json
